// apps/inatrace-fe/src/api/model/apiPlotGeoJsonImportIssue.ts
export interface ApiPlotGeoJsonImportIssue { 
    /**
     * 1-based position of the feature in the file
     */
    featureNumber?: number;
    /**
     * Farmer internal ID read from the feature, if any
     */
    farmerInternalId?: string;
    /**
     * Plot code read from the feature, if any
     */
    plotName?: string;
    /**
     * Why the feature cannot be imported
     */
    type?: ApiPlotGeoJsonImportIssue.TypeEnum;
}

/**
 * Namespace for property- and property-value-enumerations of ApiPlotGeoJsonImportIssue.
 */
export namespace ApiPlotGeoJsonImportIssue {
    /**
     * All properties of ApiPlotGeoJsonImportIssue.
     */
    export enum Properties {
        /**
         * 1-based position of the feature in the file
         */
        featureNumber = 'featureNumber',
        /**
         * Farmer internal ID read from the feature, if any
         */
        farmerInternalId = 'farmerInternalId',
        /**
         * Plot code read from the feature, if any
         */
        plotName = 'plotName',
        /**
         * Why the feature cannot be imported
         */
        type = 'type'
    }

    /**
     * All possible values of type.
     */
    export enum TypeEnum {
        MISSINGFARMERID = 'MISSING_FARMER_ID',
        FARMERNOTFOUND = 'FARMER_NOT_FOUND',
        FARMERAMBIGUOUS = 'FARMER_AMBIGUOUS',
        MISSINGPLOTNAME = 'MISSING_PLOT_NAME',
        DUPLICATEPLOTNAME = 'DUPLICATE_PLOT_NAME',
        UNSUPPORTEDGEOMETRY = 'UNSUPPORTED_GEOMETRY',
        MULTIPARTGEOMETRY = 'MULTIPART_GEOMETRY',
        POLYGONWITHHOLES = 'POLYGON_WITH_HOLES',
        INVALIDCOORDINATES = 'INVALID_COORDINATES',
        CROPNOTFOUND = 'CROP_NOT_FOUND'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiPlotGeoJsonImportIssue',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'featureNumber',
                    classname: 'ApiPlotGeoJsonImportIssue',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'farmerInternalId',
                    classname: 'ApiPlotGeoJsonImportIssue',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'plotName',
                    classname: 'ApiPlotGeoJsonImportIssue',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: true,
                    datatypeWithEnum: 'ApiPlotGeoJsonImportIssue.TypeEnum',
                    required: false,
                    name: 'type',
                    classname: 'ApiPlotGeoJsonImportIssue',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                featureNumber: [
                ],
                farmerInternalId: [
                ],
                plotName: [
                ],
                type: [
                ],
            }
        }
    }

  // export const ApiPlotGeoJsonImportIssueValidationScheme = {
  //     validators: [],
  //     fields: {
  //               featureNumber: {
  //                   validators: []
  //               },
  //               farmerInternalId: {
  //                   validators: []
  //               },
  //               plotName: {
  //                   validators: []
  //               },
  //               type: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiPlotGeoJsonImportIssue>;


}